+++ test/public-suffix/uniform-instance-destruction.js
@@ -2,11 +2,11 @@
 
 function uniformInstanceDestructionTests (t = () => {}, parentScope = [], sharedState = {}) {
   const scope = [...parentScope, uniformInstanceDestructionTests.name]
-  const {m, ps} = sharedState
+  const {jetta, m, ps} = sharedState
 
   ps.destroy()
 
-  const destroyKeys = Object.keys(ps)
+  const destroyKeys = Object.keys(jetta.PublicSuffix.keysToDestroy)
 
   for (let i = 0, len = destroyKeys.length; i < len; i++) {
     t.equal(ps[destroyKeys[i]], null, m(scope, `"${destroyKeys[i]}" should be \`null\``))