+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   - Updated logo (PNG -> SVG)
   - Added [nsp badge](https://github.com/badges/shields/pull/1156)
 - Added explicit test support for Node LTS versions
+- Fixed a bug on Node 9 where an event triggered via a destroyed `CookieManager` or `PublicSuffix` instance may throw a `Cannot read property '$property' of null` error
+  - Added static `keysToDestroy` object to each class to whitelist properties to destroy
 
 ## [1.1.3](https://github.com/AltusAero/jetta/releases/tag/v1.1.3)